Change plugin to fix IE11 cache bug
IE11 was caching the document on editors not allowing them (or users)
the new uploaded document. The header directive has been set now to
no-cache, forcing the browser to first check with the server whether the
file is new or not before serving. Plus added a legacy header Pragma
that was recommended for IE11 caching issues. Seems to work now will
have to test on a DOM1 machine.